+++ changelog.md
@@ -6,13 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 # [Unreleased]
 
+# [Released]
+
 ## [0.2.4] - 2026-08-05
 
 ### Changed
 
 - Change visibility of filter channel function (`a57bc76`)
-
-# [Released]
 
 ## [0.2.3] - 2026-08-03